delayed-inserter/requirements.txt
async-timeout==3.0.1
aioredis==1.3.0
cassandra-driver==3.18.0
elasticsearch==7.1.0
websockets==8.1
